Add templates for reminder, ready confirmation and cancelation messages / reorg template files
