+++ src/app/(master-layout)/event/[slug]/orders/order-form.tsx
@@ -162,7 +162,7 @@ export default function OrderForm({
         disabled={pending}
         className="w-full bg-slate-300 hover:bg-slate-400 "
       >
-        {pending ? "Creating..." : "Create Order"}
+        {pending ? "Ordering..." : "Create Order"}
       </Button>
     </form>
   );
