+++ src/app/webhooks/conversations/coffee-helper.ts
@@ -1,17 +1,14 @@
 import { Event } from "@/app/(master-layout)/event/[slug]/page";
 import { Order } from "@/config/menus";
+import { getCancelOrderMessage, getNoOpenOrderMessage, getOopsMessage } from "@/lib/stringTemplates";
 
-import * as templates from "@/lib/templates";
 import {
   addMessageToConversation,
   updateSyncListItem,
   updateSyncMapItem,
 } from "@/lib/twilio";
 
-const NEXT_PUBLIC_EVENTS_MAP = process.env.NEXT_PUBLIC_EVENTS_MAP || "",
-  NEXT_PUBLIC_ACTIVE_CUSTOMERS_MAP =
-    process.env.NEXT_PUBLIC_ACTIVE_CUSTOMERS_MAP || "",
-  UNLIMTED_ORDERS = (process.env.UNLIMITED_ORDERS || "").split(",");
+const NEXT_PUBLIC_EVENTS_MAP = process.env.NEXT_PUBLIC_EVENTS_MAP || "";
 
 export async function updateOrder(event: string, index: number, data: Order) {
   const item = await updateSyncListItem(event, index, data);
@@ -32,17 +29,17 @@ export async function cancelOrder(
         cancelledCount: Number(event.cancelledCount) + 1,
       });
 
-      const message = templates.getCancelOrderMessage(
+      const message = getCancelOrderMessage(
         data.item.shortTitle,
         index,
       );
       addMessageToConversation(conversationSid, message);
     } catch (error) {
-      const message = templates.getOopsMessage(error);
+      const message = getOopsMessage(error);
       addMessageToConversation(conversationSid, message);
     }
   } else {
-    const message = templates.getNoOpenOrderMessage();
+    const message = getNoOpenOrderMessage();
     addMessageToConversation(conversationSid, message);
   }
 }
